kredi_rag_sistemi/requirements_colab.txt
# Google Colab için optimize edilmiş gereksinimler
# Colab'da zaten yüklü olan paketler atlandı

# Core AI/ML libraries
llama-cpp-python>=0.2.38
huggingface_hub>=0.20.2
langchain>=0.1.4
sentence-transformers>=2.2.2
scikit-learn>=1.3.0
tqdm>=4.65.0

# Web framework (Colab için Streamlit)
streamlit>=1.28.0
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.4.2

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.2

# PDF processing
PyPDF2>=3.0.1
pdfplumber>=0.10.3
PyMuPDF>=1.23.8
pytesseract>=0.3.10
pdf2image>=1.16.3
camelot-py>=0.11.0
tabula-py>=2.7.0
easyocr>=1.7.0

# Vector databases
faiss-cpu>=1.7.4
chromadb>=0.4.18
langchain-community>=0.0.10

# Database connections (optional for Colab)
redis>=4.5.1
pymilvus>=2.3.0
pinecone-client>=2.2.1

# Security
cryptography>=41.0.0

# System monitoring
psutil>=5.9.0

# Colab specific
pyngrok>=5.1.0  # Tunnel için
google-colab  # Colab utilities 